Game Director Added
Core script added with functions that generate and shuffle a list of cards (cards are currently objects in a class declared in this script).  Cards are automatically dealt randomly to player hands (the plan is to get all of the game mechanics working before doing anything with input or visuals).  Debug buttons have been added to the inspector to perform actions which will later be done through touch input.