Core Mechanics Fully Implemented
The game is now fully playable in its most basic possible form.  The input is currently buttons in the inspector, and the output is currently the console, but as far as the Game Director goes everything's in and working as intended.  All planned mechanics have been added and tested (relatively thoroughly, although it is likely there are still a couple minor bugs).