Card Generator Implemented
Functionality to spawn and automatically populate card visuals added.  Card sprites are assigned though a scriptableObject which is passed to the CardVisualizer script.  This script references these sprites and generates a card (from a blank template) with the requested properties.  This can now be used to very easily add visuals to the game (all sprites are extremely temporary right now).